Revert "Changed to build both with cc and g++"
This reverts commit 7ed27fb098c874145c4210155330b206e97ee8aa.

Conflicts:

	Applications/OfferStreamer/Makefile

git-svn-id: https://repository.napa-wine.eu/svn/napawine-software/trunk/Applications/OfferStreamer@1585 c54fac2e-1815-407e-9b96-088d06743913

+++ chunkbuffer_helper.h
@@ -17,7 +17,7 @@ inline int cb_add_chunk(struct chunk_buffer *cb, const struct chunk *c) {
   return chbAddChunk(cb, c);
 }
 
-inline struct chunk *cb_get_chunks(const struct chunk_buffer *cb, size_t *n) {
+inline struct chunk *cb_get_chunks(const struct chunk_buffer *cb, int *n) {
  return chbGetChunks(cb, n);
 }
